requirements.txt
flask>=3.0
flask-cors>=4.0
pywebview>=5.0
requests>=2.31
google-genai>=1.0
PyMuPDF>=1.24
